refactor: minor types refactors II
