+++ src/sealSchema.ts
@@ -1,7 +1,7 @@
 import type { Merge } from 'type-fest';
 
-import type { JSONSchema, JSONSchemaObjectOutput } from './types';
 import { isObject } from './utils';
+import type { JSONSchema, JSONSchemaObjectOutput } from './utils/types';
 
 type DisableAdditionalPropertiesDeep<Value> = Value extends object
   ? Value extends { type: 'object' }
